More refactoring and documentation.
